refactor: 페이지 로드시 API 한번만 호출하도록 수정
