+++ src/app/comparetest/page.tsx
@@ -1,9 +1,38 @@
-'use client'; // 클라이언트 컴포넌트로 지정
+'use client';
 
-import CompareBar from '@/components/CompareBar/CompareBar'; // 실제 경로에 맞게 수정하세요.
+import { useState, useEffect } from 'react';
+import CompareBar from '@/components/CompareBar/CompareBar';
+import { getProductsAPI } from '@/api/products/getProductsAPI';
 import { ProductItem } from '@/types/api';
 
 const CompareTestPage = () => {
+  const [allProducts, setAllProducts] = useState<ProductItem[]>([]);
+  const [isLoading, setIsLoading] = useState(true);
+  const [error, setError] = useState<string | null>(null);
+
+  useEffect(() => {
+    const fetchAllProducts = async () => {
+      try {
+        setIsLoading(true);
+        const { list } = await getProductsAPI({ keyword: '' });
+        setAllProducts(list);
+      } catch (err) {
+        setError('상품 목록을 불러오는 데 실패했습니다.');
+        console.error(err);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+    fetchAllProducts();
+  }, []);
+
+  if (isLoading) {
+    return <div>로딩 중...</div>;
+  }
+  if (error) {
+    return <div>{error}</div>;
+  }
+
   const handleProductSelect = (product: ProductItem) => {
     console.log('Selected product:', product);
   };
@@ -17,11 +46,19 @@ const CompareTestPage = () => {
       <h1 className='text-h1-bold mb-4'>CompareBar Test</h1>
       <div className='flex gap-8'>
         <div className='flex w-[340px] items-center justify-center'>
-          <CompareBar onSelectProduct={handleProductSelect} onRemoveProduct={handleProductRemove} />
+          <CompareBar
+            products={allProducts}
+            onSelectProduct={handleProductSelect}
+            onRemoveProduct={handleProductRemove}
+          />
         </div>
 
         <div className='flex w-[340px] items-center justify-center'>
-          <CompareBar onSelectProduct={handleProductSelect} onRemoveProduct={handleProductRemove} />
+          <CompareBar
+            products={allProducts}
+            onSelectProduct={handleProductSelect}
+            onRemoveProduct={handleProductRemove}
+          />
         </div>
       </div>
     </div>
